+++ kadynsWOTRMods/Tweaks/Deities/LichDeity.cs
@@ -17,7 +17,7 @@ using System.Threading.Tasks;
 namespace kadynsWOTRMods.Tweaks.Deities {
     internal class PatchLichDeity {
 
-        private static readonly BlueprintFeature LichDeityMythicFeature = Resources.GetBlueprint<BlueprintFeature>("d633cf9ebcdc8ed4e8f2546c3e08742e");
+        
 
 
         
@@ -26,7 +26,8 @@ namespace kadynsWOTRMods.Tweaks.Deities {
                 
         public static void AddLichDeity() {
 
-                    if (ModSettings.AddedContent.Deities.IsDisabled("LichDeity")) { return; }
+                    
+                    var LichDeityMythicFeature = Resources.GetBlueprint<BlueprintFeature>("d633cf9ebcdc8ed4e8f2546c3e08742e");
                     var LichDeityFeature = Resources.GetBlueprint<BlueprintFeature>("b4153b422d02d4f48b3f8f0ceb6a10eb");
                     LichDeityFeature.RemoveComponents<PrerequisiteNoFeature>();
                     LichDeityFeature.AddComponent<AddFacts>(c => {