+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+v0.2.0
+**ALL FEATURES NOW TOGGLEABLE IN UMM SETTINGS**
+-Fixed Deskari not showing up in deity selection,
+-Added the minor deity Milani
+-Added the minor deity Ragathiel
+-Added User settings toggleable through the UMM UI
+-Fixed a spelling error in the lich deity desc
+
 *Version 0.1.0 -
 -Added a lore-friendly icon for all forbidden religions.
 -The Green Faith now has a description that includes its domains/favoured weapons!